+++ noodles-bam/examples/bam_query_async.rs
@@ -20,9 +20,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let mut reader = File::open(&src).await.map(bam::AsyncReader::new)?;
 
-    let header: sam::Header = reader.read_header().await?.parse()?;
+    let header = reader.read_header().await?.parse()?;
     let index = bai::r#async::read(src.with_extension("bam.bai")).await?;
-    let mut query = reader.query(header.reference_sequences(), &index, &region)?;
+    let mut query = reader.query(&header, &index, &region)?;
 
     let mut writer = sam::AsyncWriter::new(io::stdout());